add integration test for losing

+++ bettor/Models/BetResult.cs
@@ -3,10 +3,10 @@ namespace bettor.Models;
 public class BetResult
 {
     public string Status { get; set; }
-    public long Account { get; set; }
+    public int Account { get; set; }
     public string Points { get; set; }
 
-    public BetResult(string status, long account, string points)
+    public BetResult(string status, int account, string points)
     {
         Status = status;
         Account = account;
